+++ HISTORY.md
@@ -5,8 +5,10 @@ TODOS/WISHES
 ------------
 1. 
 
-
+0.0.3 (2023-12-14)
+-------------------
 - add BusI2c (just simple for detect_i2c_devices/*buses)
+- return only HistoryIO from write_read_line
 
 0.0.2 (2023-12-13)
 -------------------